# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /care-guides:
    get:
      operationId: listCareGuides
      tags: [careGuides]
      summary: List all care guides for the current user
      responses:
        "200":
          description: List of care guides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CareGuide"
        "401":
          description: Not signed in
    post:
      operationId: createCareGuide
      tags: [careGuides]
      summary: Generate a 30-day care guide using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCareGuideInput"
      responses:
        "201":
          description: Care guide created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareGuide"
        "400":
          description: Invalid input
        "401":
          description: Not signed in
        "503":
          description: AI generation unavailable
  /care-guides/{id}:
    get:
      operationId: getCareGuide
      tags: [careGuides]
      summary: Get a single care guide by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The care guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareGuide"
        "401":
          description: Not signed in
        "404":
          description: Not found
    delete:
      operationId: deleteCareGuide
      tags: [careGuides]
      summary: Delete a care guide
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not signed in
        "404":
          description: Not found
    patch:
      operationId: updateCareGuide
      tags: [careGuides]
      summary: Update reminder settings or completed days for a care guide
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCareGuideInput"
      responses:
        "200":
          description: Updated care guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareGuide"
        "400":
          description: Invalid input
        "401":
          description: Not signed in
        "404":
          description: Not found
  /care-guides/{id}/day30-photo:
    patch:
      operationId: uploadDay30Photo
      tags: [careGuides]
      summary: Upload the day-30 vitality photo and mark the guide as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDay30PhotoInput"
      responses:
        "200":
          description: Updated care guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareGuide"
        "400":
          description: Invalid input
        "401":
          description: Not signed in
        "404":
          description: Not found
...
components:
...
  schemas:
...
    CareGuideDailyEntry:
      type: object
      required: [day, giessen, bespruehen, beschneiden, drehen, duengen, umtopfen]
      properties:
        day:
          type: integer
          minimum: 1
          maximum: 30
        giessen:
          type: boolean
        bespruehen:
          type: boolean
        beschneiden:
          type: boolean
          nullable: true
        drehen:
          type: boolean
        duengen:
          type: boolean
        duengerart:
          type: string
          nullable: true
          description: "biologisch | mineralisch | null"
        umtopfen:
          type: boolean
          description: Repotting recommended on this day (only true on day 1 or 2 when repotting is needed)
        notizen:
          type: string
    CareGuide:
      type: object
      required: [id, userId, plantName, targetHumidity, startDate, endDate, status, dailyPlan, hasImageDay1, hasImageDay30, reminderEnabled, reminderTime, completedDays, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        plantId:
          type: integer
          nullable: true
        plantName:
          type: string
        targetHumidity:
          type: string
        potSizeRecommendation:
          type: string
          nullable: true
          description: Assessment of current pot size, e.g. "Aktueller Topf ist ausreichend" or "Umtopfen empfohlen"
        recommendedPotDiameter:
          type: string
          nullable: true
          description: Recommended minimum pot diameter in cm when repotting is needed, e.g. "18 cm"
        recommendedSoilType:
          type: string
          nullable: true
          description: Recommended soil type for this plant, e.g. "Zimmerpflanzenerde mit Perlit"
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [Aktiv, Abgeschlossen]
        dailyPlan:
          type: array
          items:
            $ref: "#/components/schemas/CareGuideDailyEntry"
        hasImageDay1:
          type: boolean
        hasImageDay30:
          type: boolean
        reminderEnabled:
          type: boolean
          description: Whether daily reminders are active for this guide.
        reminderTime:
          type: string
          description: Time of day for the reminder in HH:MM format, e.g. "09:00".
        completedDays:
          type: array
          items:
            type: integer
          description: List of day numbers (1–30) the user has marked as done.
        createdAt:
          type: string
    UpdateCareGuideInput:
      type: object
      properties:
        reminderEnabled:
          type: boolean
        reminderTime:
          type: string
          description: HH:MM format, e.g. "09:00"
        completedDays:
          type: array
          items:
            type: integer
    CreateCareGuideInput:
      type: object
      required: [plantName, healthScore, symptoms, duengeBiologisch, duegeChemisch]
      properties:
        plantId:
          type: integer
          nullable: true
        plantName:
          type: string
        imageDay1:
          type: string
          nullable: true
          description: Base64 data URL of the scan photo (Pflanzendoc day-1 image).
        healthScore:
          type: integer
          minimum: 0
          maximum: 100
        symptoms:
          type: array
          items:
            type: string
        duengeBiologisch:
          type: string
        duegeChemisch:
          type: string
    UploadDay30PhotoInput:
      type: object
      required: [imageDay30]
      properties:
        imageDay30:
          type: string
          description: Base64 data URL of the day-30 follow-up photo.
...